+++ anchor/programs/gigentic/src/contexts/mod.rs
@@ -1,6 +1,6 @@
-pub mod agent_to_consumer_rating;
-pub mod consumer_to_agent_rating;
+pub mod consumer_to_provider_rating;
 pub mod init_service;
 pub mod init_service_registry;
 pub mod pay_service;
+pub mod provider_to_consumer_rating;
 pub mod sign_service;
