+++ anchor/programs/gigentic/src/lib.rs
@@ -3,8 +3,8 @@ mod constants;
 mod contexts;
 mod errors;
 mod states;
-use contexts::agent_to_consumer_rating::*;
-use contexts::consumer_to_agent_rating::*;
+use contexts::provider_to_consumer_rating::*;
+use contexts::consumer_to_provider_rating::*;
 use contexts::init_service::*;
 use contexts::init_service_registry::*;
 use contexts::pay_service::*;
